refactor: optimize amdb
